v4.3: fix MVCC race, no-op leak, extsort cross-spill, tree compaction, dead code purge
Critical correctness fixes (proven by failing tests → fixed → passing):
1. MVCC provisional reader race: use txn_id=0 as provisional sentinel
   (blocks all reclamation) instead of MaxUint64 (which means no readers)
2. No-op commit page leak: skip chain rewrite when nothing freed/consumed
3. ExtSort last-wins across spill runs: add global seq to spill format,
   use in k-way merge for different-scope overlap resolution
4. Go scopeDirty: set in ScopeBitmapSetFeed/ClearFeed
5. Go feedBit=32: reject in bitmap mode (was silently creating scope 0)
6. Partial delete tree compaction: compact_if_needed rebuilds the tree
   when >4x sparser than needed, reusing COW copy pages for zero growth

Reader hardening:
7. Reader::open now validates CRC on meta pages (was skip_crc=true)
8. Tests updated to match CRC-validating behavior

Dead code purge (4673 lines removed):
- Rust kv.rs (1206 lines) — dead v4.1 KV metadata module, not compiled
- Rust scope.rs (482 lines) — dead v4.1 named-scope module, not compiled
- Rust metadata_conformance.rs (358 lines) — tested dead Phase 4c APIs
- Rust reader.rs meta_get/meta_list/target_kv_root — stub methods
- Rust MetaEntry type — only used by dead methods
- Rust writer.rs prev_committed_root/height — dead fields
- Rust robustness.rs: 76 dead tests for KV/named-scope APIs removed
  (2804→732 lines, 35 live tests remain)
- Go .disabled files: all 28 removed (scope.go, kv.go, fast_bytes.go, etc)
- All Phase 4c TODOs removed

Clippy: 0 errors (was 15+). SOW updated.

212 Rust + 180 Go = 392 tests pass, 0 failures, 0 warnings, 0 disabled files.

+++ v4/rust/iprange-livedb/tests/feed_bit_panic_test.rs
@@ -16,7 +16,7 @@ fn feed_bit_436_no_panic() {
     let byte_idx = 436 / 8; // 54
     let bit_idx = 436 % 8;  // 4
     assert!(bitmap.len() > byte_idx, "bitmap too short: {} bytes", bitmap.len());
-    assert!(bitmap[byte_idx as usize] & (1 << bit_idx) != 0, "feed bit 436 not set");
+    assert!(bitmap[byte_idx] & (1 << bit_idx) != 0, "feed bit 436 not set");
 }
 
 #[test]
